refactor: Extract reusable `<.itinerary_summary />` component
